Refactor `src/core/id.rs` into a module directory to resolve God Module smell.
